refactor(db-schema): split searchDocuments into its own file
searchDocuments belongs to the search domain, not AI. Move it from
schema/ai.ts to schema/search.ts; index.ts re-exports both, so
import sites are unaffected.
