+++ packages/db-schema/src/schema/index.ts
@@ -4,3 +4,4 @@ export * from './content'
 export * from './enrichment'
 export * from './migration'
 export * from './ops'
+export * from './search'
